Update .gitignore and README.md, and add Vagrantfile configuration
This commit updates the .gitignore file to exclude the .vagrant/ directory. It also updates the README.md file by adding a period at the end of the sentence and includes additional information about initializing the Vagrantfile. Lastly, it adds the configuration for the Vagrantfile, defining four virtual machines: mercurio, venus, tierra, and marte.